fix : 본문만 수정할 때 Null pointer exception 문제 해결

+++ footprint/src/main/java/com/umc/footprint/src/footprints/FootprintService.java
@@ -53,18 +53,20 @@ public class FootprintService {
             // 전달되어온 파일 리스트
             List<MultipartFile> photos = patchFootprintReq.getPhotos();
 
-            if(dbPhotoList.isEmpty()) { // 발자국에 저장된 사진이 존재하지 않음
-                if(!photos.isEmpty()) { // 전달된 파일이 하나라도 존재
-                    uploadImg(photos, userIdx, footprintIdx); // 새로운 사진들 업로드
+            if(photos != null) { // 본문만 수정(사진 수정 X)하는 경우 photos 자체가 null이 됨
+                if(dbPhotoList.isEmpty()) { // 발자국에 저장된 사진이 존재하지 않음
+                    if(!photos.isEmpty()) { // 전달된 파일이 하나라도 존재
+                        uploadImg(photos, userIdx, footprintIdx); // 새로운 사진들 업로드
+                    }
                 }
-            }
-            else { // 발자국에 저장된 기존 사진들이 존재
-                if(photos.isEmpty()) { // 전달된 파일이 없음 > 사진을 지우고 싶다는 의미 > 기존 사진 삭제만 진행
-                    footprintDao.deletePhotos(footprintIdx); // 기존 사진들 테이블에서 삭제
-                }
-                else {
-                    footprintDao.deletePhotos(footprintIdx);
-                    uploadImg(photos, userIdx, footprintIdx); // 새로운 사진들 업로드
+                else { // 발자국에 저장된 기존 사진들이 존재
+                    if(photos.isEmpty()) { // 전달된 파일이 없음 > 사진을 지우고 싶다는 의미 > 기존 사진 삭제만 진행
+                        footprintDao.deletePhotos(footprintIdx); // 기존 사진들 테이블에서 삭제
+                    }
+                    else {
+                        footprintDao.deletePhotos(footprintIdx);
+                        uploadImg(photos, userIdx, footprintIdx); // 새로운 사진들 업로드
+                    }
                 }
             }
 
@@ -92,6 +94,7 @@ public class FootprintService {
             // updateAt 업데이트
             footprintDao.updateAt(patchFootprintReq, footprintIdx);
         } catch (Exception exception) { // DB에 이상이 있는 경우 에러 메시지를 보냅니다.
+            exception.printStackTrace();
             throw new BaseException(DATABASE_ERROR);
         }
     }
